debug: only run debug spec
