+++ .github/workflows/cypress.yml
@@ -83,7 +83,7 @@ jobs:
       fail-fast: false
       matrix:
         node-version: [16]
-        containers: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11]
+        containers: [1, 2, 3, 4]
         php-versions: [ '8.1' ]
         run-in-parallel:
           - false # only for PRs: ${{ !!github.head_ref }}
